--- app/content/2.goods.yml
@@ -1,39 +1,0 @@
-title: Public Goods
-description: Here you will find all of the public goods that are made available to the Citizenry by the Stewardship, as well as goods provided to the Ecosystem by generous 3rd-parties.
-ogTitle: Public Goods - Lotusia
-ogDescription: Access the Lotusia Ecosystem
-ogImage: /img/goods_0.jpg
-to: /goods
-navigation.icon: i-mdi-fountain
-sections:
-  - title: Blockchain software - lotusd / Lotus QT
-    description: Lotus Blockchain software - Maintains a full copy of the Treasury's ledger. ** REQUIRED FOR MINERS
-    headline: '1st Samuel - 9th Epoch of the Lotus Network'
-    version: '9.4.2'
-    links:
-      - label: Windows
-        to: github/lotusd/releases/download/v9.4.2/lotus-9.4.2-win64-setup-unsigned.exe
-        icon: i-mdi-microsoft-windows
-        type: EXE Installer (.exe)
-      - label: MacOS
-        to: github/lotusd/releases/download/v9.4.2/lotus-9.4.2-osx-unsigned.dmg
-        icon: i-mdi-apple
-        type: DMG Image (.dmg)
-      - label: Linux
-        to: github/lotusd/releases/download/v9.4.2/lotus-9.4.2-x86_64-linux-gnu.tar.gz
-        icon: i-mdi-linux
-        type: TAR archive (.tar.gz)
-  - title: Prosperity Software - Lotus Vase
-    description: Flagship mobile wallet - Give and receive Lotus. More features to come!
-    links:
-      - label: Android
-        version: Coming Soon!
-        to:
-        icon: i-mdi-android
-        type: Play Store
-      - label: iOS
-        version: '2.6.1'
-        to: https://apps.apple.com/us/app/cashew-wallet/id1539306720
-        target: _blank
-        icon: i-mdi-apple
-        type: App Store
